build warning修正(#1405)
変数の型不一致を修正
デバッグコード削除

+++ TTXSamples/TTXttyrec/TTXttyrec.c
@@ -119,7 +119,7 @@ static void PASCAL TTXReadIniFile(const wchar_t *fn, PTTSet ts) {
   pvar->rec_stsize = GetOnOff(INISECTION, "RecordStartSize", fn, TRUE);
   pvar->rec_auto = GetOnOff(INISECTION, "RecordAutoStart", fn, FALSE);
 
-  size_t len = MAX_PATH;
+  DWORD len = MAX_PATH;
   wchar_t *buff = (wchar_t*)malloc(len);
   GetPrivateProfileStringW(INISECTIONW, L"RecordDefaultName", L"teraterm.tty", buff, len, fn);
   free(pvar->rec_name);
